% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPA_pdf.R
\name{ESPA_pdf}
\alias{ESPA_pdf}
\title{Computes probability density function.}
\usage{
ESPA_pdf(t, time, status)
}
\arguments{
\item{t}{A numeric. The failure time at which the density should be computed.}

\item{time}{A numeric vector. The observed time vector.}

\item{status}{A numeric vector. Indicates whether observation is censored (0) or uncensored (1)}
}
\value{
A numeric. The probability density estiamtion at t.
}
\description{
Computes probability density function.
}
