% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPA_supp.R
\name{ESPA_supp}
\alias{ESPA_supp}
\title{ESPA support}
\usage{
ESPA_supp(time, status)
}
\arguments{
\item{time}{A numeric vector. The observed time vector.}

\item{status}{A numeric vector. Indicates whether observation is censored (0) or uncensored (1)}
}
\value{
A vector of support for the distribution.
}
\description{
Computes the support for the failure time distribution.
}
