% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPA_survival.R
\name{ESPA_survival}
\alias{ESPA_survival}
\title{Computes survival function.}
\usage{
ESPA_survival(t, time, status, inc = 0.01)
}
\arguments{
\item{t}{A numeric. The failure time at which the density should be computed.}

\item{time}{A numeric vector. The observed time vector.}

\item{status}{A numeric vector. Indicates whether observation is censored (0) or uncensored (1)}

\item{inc}{A numeric. What should be the increment to be used to plot the survival function.}
}
\value{
A numeric. The survival function estiamtion at t.
}
\description{
Computes survival function.
}
